# Update Random with faster implementation + some clean up