This replaces the current Mersenne Twister with a new pseudorandom number generator known as Small Fast Counter 32. SFC32 seemed to be the best all around PRNG that I could find that worked well across the targets I can test on (HTML5 Electron and Firefox, Windows...yeah still limited here, sorry). 

SFC32 was always faster or generated more numbers in a shorter time period than the current implementation in a few various test cases that I tried out which are all obviously contrived and more about trying to identify a noticeable difference rather than trying to extract an exact value. Those cases were the following:

1. Out of a list of 50k entities, select an entity and create a new Vector2 with a new random X and new random Y float position for it
2. Generate an array of 50k ints
3. Generate as many ints as possible in 10 seconds

For case 1 and 2, this was done over a period of 10 seconds at 60 updates per second. This was using a timetask, however these tests weren't stressful enough to trigger a slowdown which would mess with the final timings (I made sure to check for that). For case 3, I basically just ended up getting bottlenecked somewhere across all targets for both implementations. Dunno why, don't really care because SFC32 still wins.

For every case, the values are rough and eyeballed (laziness), but still show a trend.

Test Case 1 (Vectors): Cumulative time spent
| Impl | Windows | Electron | Firefox |
|---|---|---|---|
| SFC32 | 0.382s | 0.997s | 0.999s |
| MT | 0.952s | 1.396s | 1.757s | 

Test Case 2 (Int Array): Cumulative time spent